requirements.txt
openai
azure.cognitiveservices.speech
threading